Fixed transient stacking and windows cycling.
Transients are usually opened form the active client but it is not
mandatory in order to have transients properly stacked we stack
them just in front of the client they are transient for.

ActivateNext was using LowerClient to change the state (from above to below)
that's not the expected behavior as LowerClient change the stacking order we just want to raise the next client and change the previous one state.